Refactored registry and mirror
